requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.12
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jsonapi==2.2.0
djangorestframework-jwt==1.11.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==19.7.1
idna==2.6
inflection==0.3.1
iso8601==0.1.12
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
promise==2.1
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
Rx==1.6.0
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==1.0.2
